requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
pg8000==1.31.2          # Cloud SQL PostgreSQL driver (pure Python, no native libs needed)
# Uncomment below if using Cloud SQL connector
# cloud-sql-python-connector[pg8000]==1.9.0

# Google AI (Gemini)
google-generativeai==0.7.2

# Google Calendar / OAuth
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utilities
python-dotenv==1.0.1